requirements.txt
fastapi==0.115.5
pydantic==2.9.2
PyMySQL==1.1.1
python-dotenv==1.0.1
uvicorn==0.32.1
langchain==0.3.9
huggingface-hub==0.26.3
